Split client and server SIWE code
