+++ packages/connectkit-next-siwe/src/index.ts
@@ -1,1 +1,2 @@
-export { configureClientSIWE, configureServerSideSIWE } from './configureSIWE';
+export { configureClientSIWE } from './client'
+export { configureServerSideSIWE } from './configureSIWE';
